1.Version
Some features have not been implemented yet. But the core calculator and its mathematical (self-written) algorithms are up and running. The Interface is in German but it is easy to figure out if one is familiar with binomial calculations. The Code is well enough commented and mostly in English. The whole app is created using QT Creator, so one might need that to work on the project.